+++ Sources/LaunchingService/Private/NoticeChecker.swift
@@ -5,23 +5,18 @@
 //  Created by SwiftMan on 2023/02/11.
 //
 
+import Dependencies
 import Foundation
 
 struct NoticeChecker: Sendable {
-  private let dateProvider: any DateProviding
-
-  init(dateProvider: any DateProviding = SystemDateProvider()) {
-    self.dateProvider = dateProvider
-  }
+  @Dependency(\.date.now)
+  private var now
 
   func compare(launching: Launching) -> AppUpdateStatus {
-    if let notice = launching.notice, notice.dateRange.contains(dateProvider.now) {
-      return .notice(NoticeAlert(title: notice.title,
-                                 message: notice.message,
-                                 isAppTerminated: notice.isAppTerminated,
-                                 doneURL: notice.doneURL))
+    if let notice = launching.notice, notice.dateRange.contains(now) {
+      return .notice(NoticeAlert(notice: notice))
     }
-    
+
     return .valid
   }
 }